move mocks to separate package
